backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-ckeditor-5==0.2.11
Pillow==10.4.0
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg[binary]==3.2.13
djangorestframework-simplejwt==5.3.0 
